route accessing fixed
